9 optimize mctd mc laser

+++ cat_laser_roi/views.py
@@ -49,12 +49,11 @@ def run_optimization(request):
             pieces_data = data.get('pieces_data')
 
             piece_names = [str(item[0]) for item in pieces_data if item and item[0] is not None]
-            piece_lengths = [int(item[1]) for item in pieces_data if item and item[1] is not None]
+            
+            piece_lengths = [float(item[1]) for item in pieces_data if item and item[1] is not None]
             demands_list = [int(item[2]) for item in pieces_data if item and item[2] is not None]
             priorities_list = [int(item[3]) for item in pieces_data if item and item[3] is not None]
 
-            # Giới hạn 10mm, hao hụt 1.5% của 1 chiều dài cây sắt
-            # chiều dài cây sắt, tổng số lượng các đoạn, tia laser, hao hụt 1.5%, bắt đầu hao hụt 10mm, hao hut trim_start=10 + >=0
             patterns_data = get_or_calculate_patterns(
                 stock_length, piece_lengths, 1, 0.015, 10, 0
             )
